requirements.txt
absl-py==2.1.0
altgraph==0.17.4
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
flatbuffers==2.0.7
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
numpy==1.24.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
protobuf==3.19.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
pywin32-ctypes==0.2.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.18.2
